requirements.txt
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.3
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.3
openai==2.30.0
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.1.0
reportlab==4.4.10
requests==2.33.0
scikit-learn==1.8.0
scipy==1.17.1
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
